Finished Tag, GiftCertificates, Orders, Users entites, daos, services, controllers. Added some exceptions. Still need to add tests.

+++ service/src/main/java/com/epam/esm/service/TagService.java
@@ -2,63 +2,77 @@ package com.epam.esm.service;
 
 import com.epam.esm.dao.TagDao;
 import com.epam.esm.dto.TagDto;
+import com.epam.esm.exception.WrongPageException;
+import com.epam.esm.model.Page;
 import com.epam.esm.model.Tag;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.Locale;
 import java.util.Objects;
-import java.util.Set;
 import java.util.stream.Collectors;
 
 @Service
 public class TagService {
 
     private final TagDao tagDao;
+    private final ObjectConverter objectConverter;
 
     @Autowired
-    public TagService(TagDao tagDao) {
+    public TagService(TagDao tagDao, ObjectConverter objectConverter) {
         this.tagDao = Objects.requireNonNull(tagDao);
+        this.objectConverter = Objects.requireNonNull(objectConverter);
     }
 
     public TagDto saveTag(TagDto tagDto) {
-        Objects.requireNonNull(tagDto.getName());
-        return
-                convertTagToTagDto(
-                        tagDao.save(
-                                convertTagDtoToTag(tagDto)
-                        )
-                );
-    }
-
-    public TagDto selectTagByNameOrId(TagDto tagDto) {
-        if (tagDto.getId() != null) {
-            return convertTagToTagDto(tagDao.findById(tagDto.getId()));
-        } else if (tagDto.getName() != null) {
-            return convertTagToTagDto(tagDao.findByName(tagDto.getName().toLowerCase(Locale.ROOT)));
+        if (tagDto.getName() == null || tagDto.getName().isEmpty()) {
+            throw new NullPointerException("Field name cannot be empty");
         }
-        throw new NullPointerException("id or name can't be null");
+
+        return objectConverter.convertTagToTagDto(
+                tagDao.save(
+                        objectConverter.convertTagDtoToTag(tagDto)));
     }
 
-    public Set<TagDto> selectAllTags() {
-        return tagDao.findAll().stream().map(TagService::convertTagToTagDto).collect(Collectors.toSet());
+    public TagDto selectTagById(Long id) {
+        if (id == null) {
+            throw new NullPointerException("Id cannot be null");
+        }
+
+        return objectConverter
+                .convertTagToTagDto(
+                        tagDao.findById(id));
+    }
+
+    public TagDto selectTagByName(String name) {
+        if (name == null || name.isEmpty()) {
+            throw new NullPointerException("Name cannot be null or empty");
+        }
+        return objectConverter
+                .convertTagToTagDto(
+                        tagDao.findByName(
+                                name.toLowerCase(Locale.ROOT)));
+
+    }
+
+    public Page<TagDto> selectPageOfTags(Integer page) {
+        if (page < 1) throw new WrongPageException("Page cannot be smaller by 1");
+        Page<Tag> tagPage = tagDao.findPage(page);
+        return new Page<>(
+                tagPage.getPageNumber(),
+                tagPage.getPageSize(),
+                tagPage.getTotalPages(),
+                tagPage.getCollection()
+                        .stream()
+                        .map(objectConverter::convertTagToTagDto)
+                        .collect(Collectors.toList()));
     }
 
     public void deleteTag(Long id) {
+        if (id == null) {
+            throw new NullPointerException("Id cannot be null");
+        }
+
         tagDao.removeById(id);
-    }
-
-    static TagDto convertTagToTagDto(Tag tag) {
-        return new TagDto(
-                tag.getId(),
-                tag.getName()
-        );
-    }
-
-    static Tag convertTagDtoToTag(TagDto tagDto) {
-        Tag tag = new Tag();
-        tag.setId(tagDto.getId());
-        tag.setName(tagDto.getName().toLowerCase(Locale.ROOT));
-        return tag;
     }
 }
